# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel SIP Swagger API Documentation",
        "description": "Implementation of Swagger with in Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Url from config file"
        }
    ],
    "paths": {
...
        "/api/product": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get all product",
                "description": "get all product",
                "operationId": "all-product",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Anauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create product",
                "description": "create product",
                "operationId": "create-product",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "description": "Category id",
                                        "type": "integer"
                                    },
                                    "name": {
                                        "description": "Name",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Price",
                                        "type": "integer"
                                    },
                                    "sku": {
                                        "description": "SKU",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Status",
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ]
                                    },
                                    "image": {
                                        "description": "Image",
                                        "type": "file"
                                    },
                                    "description": {
                                        "description": "Description",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "401": {
                        "description": "Anauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/product/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Detail product",
                "description": "get detail product",
                "operationId": "show-product",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "Anauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete product",
                "description": "Delete product",
                "operationId": "delete-product",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "401": {
                        "description": "Anauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/product/{id}/update": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Update product",
                "description": "update product",
                "operationId": "update-product",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "description": "Category id",
                                        "type": "integer"
                                    },
                                    "name": {
                                        "description": "Name",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Price",
                                        "type": "integer"
                                    },
                                    "sku": {
                                        "description": "SKU",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Status",
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ]
                                    },
                                    "image": {
                                        "description": "Image",
                                        "type": "file"
                                    },
                                    "description": {
                                        "description": "Description",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    },
                    "401": {
                        "description": "Anauthorized"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}